mdx page template
