+++ package.json
@@ -18,7 +18,10 @@
 		"prepare": "husky install"
 	},
 	"dependencies": {
+		"@mdx-js/loader": "^2.3.0",
+		"@mdx-js/react": "^2.3.0",
 		"@next/font": "^13.0.5",
+		"@next/mdx": "^13.2.3",
 		"@nivo/circle-packing": "^0.80.0",
 		"@nivo/core": "^0.80.0",
 		"@testing-library/jest-dom": "^5.16.5",
